# oj-fronted/OPENAPI/Doj.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Doj",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/question/add": {
      "post": {
        "summary": "创建",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionAddRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseLong",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": 0,
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/delete": {
      "post": {
        "summary": "删除",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": false,
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/update": {
      "post": {
        "summary": "更新（仅管理员)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdateRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": false,
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/get": {
      "get": {
        "summary": "根据 id 获取(ADMIN)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseQuestionAdminVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": {
                        "id": 0,
                        "title": "",
                        "content": "",
                        "tags": [
                          ""
                        ],
                        "answer": "",
                        "submitNum": 0,
                        "acceptedNum": 0,
                        "judgeConfig": {
                          "timeLimit": 0,
                          "memoryLimit": 0,
                          "stackLimit": 0
                        },
                        "judgeCase": "",
                        "thumbNum": 0,
                        "favourNum": 0,
                        "userId": 0,
                        "createTime": "",
                        "updateTime": "",
                        "userVO": {
                          "id": 0,
                          "userName": "",
                          "userAvatar": "",
                          "userProfile": "",
                          "userRole": "",
                          "createTime": ""
                        }
                      },
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/get/vo": {
      "get": {
        "summary": "根据 id 获取(脱敏)",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseQuestionVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": {
                        "id": 0,
                        "title": "",
                        "content": "",
                        "tags": [
                          ""
                        ],
                        "submitNum": 0,
                        "acceptedNum": 0,
                        "judgeConfig": {
                          "timeLimit": 0,
                          "memoryLimit": 0,
                          "stackLimit": 0
                        },
                        "thumbNum": 0,
                        "favourNum": 0,
                        "userId": 0,
                        "createTime": "",
                        "updateTime": "",
                        "userVO": {
                          "id": 0,
                          "userName": "",
                          "userAvatar": "",
                          "userProfile": "",
                          "userRole": "",
                          "createTime": ""
                        }
                      },
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/list/page/vo": {
      "post": {
        "summary": "分页获取列表（封装类）",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageQuestionVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": {
                        "records": [
                          {
                            "id": 0,
                            "title": "",
                            "content": "",
                            "tags": [
                              ""
                            ],
                            "submitNum": 0,
                            "acceptedNum": 0,
                            "judgeConfig": {
                              "timeLimit": 0,
                              "memoryLimit": 0,
                              "stackLimit": 0
                            },
                            "thumbNum": 0,
                            "favourNum": 0,
                            "userId": 0,
                            "createTime": "",
                            "updateTime": "",
                            "userVO": {
                              "id": 0,
                              "userName": "",
                              "userAvatar": "",
                              "userProfile": "",
                              "userRole": "",
                              "createTime": ""
                            }
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "orders": [
                          {
                            "column": "",
                            "asc": false
                          }
                        ],
                        "optimizeCountSql": false,
                        "searchCount": false,
                        "optimizeJoinOfCountSql": false,
                        "maxLimit": 0,
                        "countId": ""
                      },
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/my/list/page/vo": {
      "post": {
        "summary": "分页获取当前用户创建的资源列表",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageQuestionVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": {
                        "records": [
                          {
                            "id": 0,
                            "title": "",
                            "content": "",
                            "tags": [
                              ""
                            ],
                            "submitNum": 0,
                            "acceptedNum": 0,
                            "judgeConfig": {
                              "timeLimit": 0,
                              "memoryLimit": 0,
                              "stackLimit": 0
                            },
                            "thumbNum": 0,
                            "favourNum": 0,
                            "userId": 0,
                            "createTime": "",
                            "updateTime": "",
                            "userVO": {
                              "id": 0,
                              "userName": "",
                              "userAvatar": "",
                              "userProfile": "",
                              "userRole": "",
                              "createTime": ""
                            }
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "orders": [
                          {
                            "column": "",
                            "asc": false
                          }
                        ],
                        "optimizeCountSql": false,
                        "searchCount": false,
                        "optimizeJoinOfCountSql": false,
                        "maxLimit": 0,
                        "countId": ""
                      },
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/list/page": {
      "post": {
        "summary": "分页获取题目列表（仅管理员）",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageQuestion",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": {
                        "records": [
                          {
                            "id": 0,
                            "title": "",
                            "content": "",
                            "tags": "",
                            "answer": "",
                            "submitNum": 0,
                            "acceptedNum": 0,
                            "judgeCase": "",
                            "judgeConfig": "",
                            "thumbNum": 0,
                            "favourNum": 0,
                            "userId": 0,
                            "createTime": "",
                            "updateTime": "",
                            "isDelete": 0
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "orders": [
                          {
                            "column": "",
                            "asc": false
                          }
                        ],
                        "optimizeCountSql": false,
                        "searchCount": false,
                        "optimizeJoinOfCountSql": false,
                        "maxLimit": 0,
                        "countId": ""
                      },
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/edit": {
      "post": {
        "summary": "编辑（用户）",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionEditRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseBoolean",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": false,
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/question/question_submit/do": {
      "post": {
        "summary": "提交题目",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionSubmitAddRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseLong",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "data": 0,
                      "message": ""
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "BaseResponseLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "data": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "message": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "JudgeCase": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string",
            "description": "输入用例"
          },
          "output": {
            "type": "string",
            "description": "输出用例"
          }
        }
      },
...
      "QuestionAddRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "标题"
          },
          "content": {
            "type": "string",
            "description": "内容"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签列表"
          },
          "answer": {
            "type": "string",
            "description": "题目答案"
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase",
              "description": "题目用例"
            },
            "description": "判题用例"
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig",
            "description": "判题配置"
          }
        }
      },
...
      "BaseResponseBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "data": {
            "type": "boolean",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "QuestionUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "id",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "description": "标题"
          },
          "content": {
            "type": "string",
            "description": "内容"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "标签列表"
          },
          "answer": {
            "type": "string",
            "description": "题目答案"
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase",
              "description": "题目用例"
            },
            "description": "判题用例"
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig",
            "description": "判题配置"
          }
        }
      },
...
      "DeleteRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "id",
            "format": "int64"
          }
        }
      },
...
      "BaseResponseQuestionVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/QuestionVO",
            "description": "对应题目信息"
          },
          "message": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "BaseResponseQuestionAdminVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/QuestionAdminVO",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "BaseResponsePageQuestionVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PageQuestionVO",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "BaseResponsePageQuestion": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PageQuestion",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api",
      "description": "开发环境"
    }
  ]
}